+++ cli/src/commands/squash.rs
@@ -26,6 +26,8 @@ use jj_lib::matchers::Matcher;
 use jj_lib::merge::Diff;
 use jj_lib::object_id::ObjectId as _;
 use jj_lib::repo::Repo as _;
+use jj_lib::revset;
+use jj_lib::revset::parse_program;
 use jj_lib::rewrite;
 use jj_lib::rewrite::CommitWithSelection;
 use jj_lib::rewrite::merge_commit_trees;
@@ -38,6 +40,7 @@ use crate::cli_util::WorkspaceCommandTransaction;
 use crate::cli_util::compute_commit_location;
 use crate::cli_util::print_unmatched_explicit_paths;
 use crate::command_error::CommandError;
+use crate::command_error::print_parse_diagnostics;
 use crate::command_error::user_error;
 use crate::complete;
 use crate::description_util::add_trailers;
@@ -46,6 +49,7 @@ use crate::description_util::description_template;
 use crate::description_util::edit_description;
 use crate::description_util::join_message_paragraphs;
 use crate::description_util::try_combine_messages;
+use crate::revset_util;
 use crate::ui::Ui;
 
 /// Move changes from a revision into another revision
@@ -53,12 +57,8 @@ use crate::ui::Ui;
 /// Without any options, moves the changes from the working-copy revision to the
 /// parent revision.
 ///
-/// With the `-r` option, moves the changes from the specified revision to the
-/// parent revision. Fails if there are several parent revisions (i.e., the
-/// given revision is a merge).
-///
-/// With the `--from` and/or `--into` options, moves changes from/to the given
-/// revisions. If either is left out, it defaults to the working-copy commit.
+/// Moves changes from `--from` argument(s) to the `--into` argument.
+/// If `--into` is not specified, it defaults to `heads(from)-`.
 /// For example, `jj squash --into @--` moves changes from the working-copy
 /// commit to the grandparent.
 ///
@@ -84,25 +84,19 @@ use crate::ui::Ui;
 /// `--from @` is assumed).
 #[derive(clap::Args, Clone, Debug)]
 pub(crate) struct SquashArgs {
-    /// Revision to squash into its parent (default: @). Incompatible with the
-    /// experimental `-o`/`-A`/`-B` options.
-    #[arg(long, short, value_name = "REVSET")]
-    #[arg(add = ArgValueCompleter::new(complete::revset_expression_mutable))]
-    revision: Option<RevisionArg>,
-
     /// Revision(s) to squash from (default: @)
-    #[arg(long, short, conflicts_with = "revision", value_name = "REVSETS")]
-    #[arg(add = ArgValueCompleter::new(complete::revset_expression_mutable))]
-    from: Vec<RevisionArg>,
-
-    /// Revision to squash into (default: @)
     #[arg(
         long,
-        short = 't',
-        conflicts_with = "revision",
-        visible_alias = "to",
-        value_name = "REVSET"
+        short,
+        visible_alias = "from",
+        visible_short_alias = 'f',
+        value_name = "REVSETS"
     )]
+    #[arg(add = ArgValueCompleter::new(complete::revset_expression_mutable))]
+    revisions: Vec<RevisionArg>,
+
+    /// Revision to squash into (default: `heads(from)-`)
+    #[arg(long, short = 't', visible_alias = "to", value_name = "REVSET")]
     #[arg(add = ArgValueCompleter::new(complete::revset_expression_mutable))]
     into: Option<RevisionArg>,
 
@@ -113,7 +107,7 @@ pub(crate) struct SquashArgs {
         visible_alias = "destination",
         short,
         visible_short_alias = 'd',
-        conflicts_with_all = ["into", "revision"],
+        conflicts_with = "into",
         value_name = "REVSETS"
     )]
     #[arg(add = ArgValueCompleter::new(complete::revset_expression_all))]
@@ -125,7 +119,7 @@ pub(crate) struct SquashArgs {
         long,
         short = 'A',
         visible_alias = "after",
-        conflicts_with_all = ["onto", "into", "revision"],
+        conflicts_with_all = ["onto", "into"],
         value_name = "REVSETS"
     )]
     #[arg(add = ArgValueCompleter::new(complete::revset_expression_all))]
@@ -137,7 +131,7 @@ pub(crate) struct SquashArgs {
         long,
         short = 'B',
         visible_alias = "before",
-        conflicts_with_all = ["onto", "into", "revision"],
+        conflicts_with_all = ["onto", "into"],
         value_name = "REVSETS"
     )]
     #[arg(add = ArgValueCompleter::new(complete::revset_expression_mutable))]
@@ -189,46 +183,58 @@ pub(crate) async fn cmd_squash(
 
     let mut workspace_command = command.workspace_helper(ui).await?;
 
-    let mut sources: Vec<Commit>;
+    let mut sources: Vec<Commit> = if args.revisions.is_empty() {
+        workspace_command.parse_revset(ui, &RevisionArg::AT)?
+    } else {
+        workspace_command.parse_union_revsets(ui, &args.revisions)?
+    }
+    .evaluate_to_commits()?
+    .try_collect()
+    .await?;
+
+    let sources_str: &str = if sources.is_empty() {
+        "none()"
+    } else {
+        &sources
+            .iter()
+            .map(|c| c.id().hex())
+            .collect::<Vec<_>>()
+            .join("|")
+    };
+
     let pre_existing_destination;
 
-    if !args.from.is_empty() || args.into.is_some() || insert_destination_commit {
-        sources = if args.from.is_empty() {
-            workspace_command.parse_revset(ui, &RevisionArg::AT)?
-        } else {
-            workspace_command.parse_union_revsets(ui, &args.from)?
-        }
-        .evaluate_to_commits()?
-        .try_collect()
-        .await?;
-        if insert_destination_commit {
-            pre_existing_destination = None;
-        } else {
-            let destination = workspace_command
-                .resolve_single_rev(ui, args.into.as_ref().unwrap_or(&RevisionArg::AT))
-                .await?;
-            // remove the destination from the sources
-            sources.retain(|source| source.id() != destination.id());
-            pre_existing_destination = Some(destination);
-        }
-        // Reverse the set so we apply the oldest commits first. It shouldn't affect the
-        // result, but it avoids creating transient conflicts and is therefore probably
-        // a little faster.
-        sources.reverse();
+    if insert_destination_commit {
+        pre_existing_destination = None;
     } else {
-        let source = workspace_command
-            .resolve_single_rev(ui, args.revision.as_ref().unwrap_or(&RevisionArg::AT))
+        let into_str = args
+            .into
+            .as_ref()
+            .map(|arg| arg.as_ref())
+            .unwrap_or("heads(from)-");
+
+        let mut context = workspace_command.env().revset_parse_context();
+        context
+            .local_variables
+            .insert("from", parse_program(sources_str)?);
+        let mut diags = revset::RevsetDiagnostics::default();
+        let expression = revset::parse(&mut diags, into_str, &context)?;
+        print_parse_diagnostics(ui, "In --into", &diags)?;
+
+        let evaluator = workspace_command.attach_revset_evaluator(expression);
+        let destination =
+            revset_util::evaluate_revset_to_single_commit(into_str, &evaluator, || {
+                workspace_command.commit_summary_template()
+            })
             .await?;
-        let mut parents = source.parents().await?;
-        if parents.len() != 1 {
-            return Err(
-                user_error("Cannot squash merge commits without a specified destination")
-                    .hinted("Use `--into` to specify which parent to squash into"),
-            );
-        }
-        sources = vec![source];
-        pre_existing_destination = Some(parents.pop().unwrap());
+        // remove the destination from the sources
+        sources.retain(|source| source.id() != destination.id());
+        pre_existing_destination = Some(destination);
     }
+    // Reverse the set so we apply the oldest commits first. It shouldn't affect the
+    // result, but it avoids creating transient conflicts and is therefore probably
+    // a little faster.
+    sources.reverse();
 
     workspace_command
         .check_rewritable(sources.iter().chain(&pre_existing_destination).ids())
@@ -419,7 +425,7 @@ pub(crate) async fn cmd_squash(
         }
 
         if let [only_path] = &*args.paths {
-            let no_rev_arg = args.revision.is_none() && args.from.is_empty() && args.into.is_none();
+            let no_rev_arg = args.revisions.is_empty() && args.into.is_none();
             if no_rev_arg
                 && tx
                     .base_workspace_helper()